refactor: moving stuff in vote to own ui classes
