+++ mobile-app/lib/src/vote/ui/votes_page.dart
@@ -41,101 +41,4 @@ class VotesPage extends StatelessWidget {
       ),
     );
   }
-
-  Widget _buildVoteList(List<Vote> votes) {
-    return ListView.builder(
-      itemCount: votes.length,
-      itemBuilder: (context, index) {
-        final vote = votes[index];
-        return _buildVoteCard(
-          context,
-          vote,
-        );
-      },
-    );
-  }
-
-  String __formatLocation(Location? location) {
-    return location != null
-        ? [
-            location.city,
-            location.address != null ? location.address!.street : "",
-            location.country,
-          ].where((element) => element.isNotEmpty).join(", ")
-        : "No location available";
-  }
-
-  Widget _buildVoteCard(BuildContext context, Vote vote) {
-    var optionCounts = vote.options.map((option) => option.votesCount).toList();
-    var totalVotes = optionCounts.reduce((a, b) => a + b);
-
-    return Card(
-      margin: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          ListTile(
-            leading: CircleAvatar(
-              backgroundImage: vote.event.imageUrl == null
-                  ? AssetImage('assets/image_not_found.png')
-                  : NetworkImage(vote.event.imageUrl!),
-            ),
-            title: Text(vote.event.title,
-                style: TextStyle(fontWeight: FontWeight.bold)),
-            subtitle: Text(__formatLocation(vote.event.location)),
-            trailing: Icon(Icons.star_border),
-            onTap: () {
-              context.push('${AppRoute.events.path}/${vote.event.id}',
-                  extra: vote.event);
-            },
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(vote.text, style: TextStyle(fontWeight: FontWeight.bold)),
-                const SizedBox(height: 8),
-                ...vote.options.map((option) =>
-                    _buildOptionCard(context, option, totalVotes, vote)),
-              ],
-            ),
-          ),
-          const SizedBox(height: 12),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildOptionCard(
-      BuildContext context, VoteOption option, int totalVotes, Vote vote) {
-    return Card(
-      margin: const EdgeInsets.symmetric(vertical: 4),
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-      child: ListTile(
-        leading: CircleAvatar(
-          child: IconButton(
-            icon: option.isSelected
-                ? Icon(Icons.circle)
-                : Icon(Icons.circle_outlined),
-            onPressed: () {
-              context
-                  .read<VoteBloc>()
-                  .add(VoteOnOption(vote.id, option.id, !option.isSelected));
-            },
-          ),
-        ),
-        title: Text(option.text),
-        subtitle: LinearProgressIndicator(
-          value: option.votesCount / totalVotes,
-          minHeight: 6,
-          backgroundColor: Colors.grey[200],
-          color: Colors.blue,
-        ),
-        trailing: Text(
-            '${(option.votesCount / totalVotes * 100).toStringAsFixed(2)}%'),
-      ),
-    );
-  }
 }
